requirements.txt
Adafruit-ADS1x15==1.0.2
Adafruit-GPIO==1.0.3
Adafruit-PureIO==1.0.4
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
dbus-python==1.2.16
decorator==4.4.1
evento==1.0.2
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.17.0
idna==2.9
mido==1.2.9
oauthlib==3.1.0
omxplayer-wrapper==0.3.2
pathlib2==2.3.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyserial==3.4
python-rtmidi==1.4.2
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
RPi.GPIO==0.7.1
rsa==4.0
six==1.14.0
sixfab-cellulariot==1.2.0
smbus==1.1.post2
spidev==3.4
uritemplate==3.0.1
urllib3==1.25.8
